% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitstring.R
\name{splitstring}
\alias{splitstring}
\title{Split a string into a vector of strings}
\usage{
splitstring(x, sep = ",")
}
\arguments{
\item{x}{a character string to be split}

\item{sep}{the separator on which to split}
}
\value{
a vector of strings
}
\description{
Split a string into a vector of strings
}
\examples{
splitstring("a,b,c")
splitstring("alpha,bravo,charlie")
}
